optimize lcd spi
- optionally to use fast spi pins
- move lcd code to implementation
- direct register access (no driver required)
- optimize screen update to take least possible time (result ~ 30fps)
