+++ components/driver/Kconfig
@@ -1,0 +1,26 @@
+menu "LCD SPI Pins"
+
+config LCD_USE_FAST_PINS
+   bool "Use fast SPI pins"
+   default "n"
+   help
+      When enabled LCD pins are:
+          MISO 19
+          MOSI 23
+          CLK  18
+          CS   5
+          DC   21
+          RST  22
+          BCKL 25
+
+      When disabled (default) LCD pins are:
+          MISO 25
+          MOSI 23
+          CLK  19
+          CS   22
+          DC   21
+          RST  18
+          BCKL 5
+
+
+endmenu
